preserveTempPath, script fragments ForceDirectories
[frontend]
- New hidden setting to preserve tempPath.  Can be set in the ini file
directory.
- Now doing ForceDirectories for the destination path when copying
script fragments.

+++ frontend/mpMergeForm.pas
@@ -462,7 +462,8 @@ begin
   SaveSettings;
 
   // delete temppath
-  DeleteDirectory(TempPath);
+  if not settings.preserveTempPath then
+    DeleteDirectory(TempPath);
   Action := caFree;
 
   // free progress form
